Fixed handling of undo-tree-mode new version 0.7.4 released in 2020-Jan-19.
The new version does not allow using the undo-tree-undo when the undo-tree-mode is not active,
as the previous version did.
So PEL must now activate the global undo-tree mode when the pel-use-undo-tree is t.
Because of that PEL now also needs to deal with the possibility that the user toggles the mode
on/off either locally or globally.  The new code handles that by using a new pel-undo file that
defined pel-undo and pel-redo functions which check the undo-tree modes to determine how to behave.
The loading is done in 3 stages with both pel-undo and undo-tree files auto-loaded but only when
pel-use-undo-tree is t.  If it is false, nothing is loaded and the Emacs default undo command is the
only one available.  The new code also prevents the undo-tree modes to change the binding of C-_ and M-_
keys which PEL uses for negative-argument to help increase the typing flow: the code now updates the
undo-tree key-map.