Add centimacro support. With bug fixes I submitted.
Currently the version of centimacro that will be installed by PEL is a copy of my modified
version of centimacro taken from my fork of centimacro.  This should be temporary: I fixed
bugs in centimacro and sent a pull-request to abo-abo/centimacro (the original author of
centimacro).  Once the pull-request is accepted and centimacro is updated in MELPA I will
then remove that and PEL will just use the MELPA version.  In the mean time PEL will download
my copy which solves a bug in centi-restore-all.