Change max-specpdl-size to 2000 while running benchmark-init commands.
With a relatively large set of packages (170) the possibility that some package
has a large number of undwind call gets high and the normal limit of 1000 may be
too low preventing benchmark-init to complete.  Increasing the limit during
benchmark-init execution seem to prevent the issue I had when I installed treemacs.