New pel-cc-newline command assigned to RET.  This command can indent, align according to several user options and variables, allowing dynamic behaviour change with several other commands.
Required code refactoring: a new pel-align.el file holds code that is also used by pel-cc.el which holds the new function and the function that displays the state of all options.
Updated the PDF files, but still have to update the one for C++ and D, which are affected by the change.