Change PEL key binding to xref-find-apropos.
More to come, as I'm trying to create an environment where cross-reference
will "always work" with the same major keys (``M-.`` and ``<f11> X .``)
picking what is most appropriate for the major-mode and the project.  Creating
this will take some time as I'm learning the intricacies of the various cross
referencing packages, their abilities and limitations.