Change PEL key binding of xref-etags-mode to ``<f11> X B T`` using 'T' for 'T'AGS.
Ideally I would like to find a way to activate the elisp xref-backend and
would assign it to ``<f11> X B E`` that was previously used by
xref-etags-mode, and would modify the other key bindings to make them select
the mode instead of toggling them in some case.  Again all of this would be
done in the perspective to have a simple cross referencing system that would
always work, picking up the technology that works best or trying several,
while leaving the ability so select something else via some keys.