Add pel-package-install; same as package-install but if the installation fails once it tries again after refreshing the package description list.
That handles failures occurring when the package list in the computer is old enough so that the full
package name (which includes a version) corresponds to a version that is no longer held in the selected
Elpa archive.  Refreshing the list with the very latest will get the latest names with their versions and then
the next install attempt will get the latest version.
This issue affects use-package :ensure t and is causing installation failures.
This may be confusing for new Emacs users as they may not know to manually refresh
the package list to solve the problem. Therefore this new function that I will start using in several places.
For now it is only used by pel-require.