Major refactoring: replaced using use-package by local logic.
The new code can be byte-compiled in the absence of all external packages, something use-package does not allow
because use-package is meant to be placed in the init.el file and assumes that you want to download the packages
as soon as you have placed the use-package statement.   PEL is different: the goal is to be able to control
the download and activation of external packages through customization and allow all code to be byte-compiled
with all warnings fixed.
This new commit provides what is needed to byte-compile all PEL code when no external package is installed.
There is one exception: the PEL hydra code cannot be byte-compiled in the absence of the Hydra package.
To solve this, the PEL Hydra code was moved into the pel__hydra.el file.  That file is byte-compiled, but
only when pel-use-hydra is turned on and then the Hydra package is installed.  The Makefile does not control
its compilation; the code inside pel_keys.el controls the byte-compilation of pel__hydra.el.

I have not yet updated the documentation.  To do next.