Refactored pel-goto-symbol and its equivalent for all buffers.
Code now uses lexical binding and supports more UI methods to select the imenu-based target
in the current buffer or over all current buffers.
More refactoring in the way keys are bound is coming, trying to use a more general method for the
dynamic selection of configuration and providing access to related PEL and package customization.
The dynamic configuration will probably be a key sequence that ends with F4, similar to what is used
for PEL customization (ends with F2) or package customization (ends with F3).
Will also probably look for more consolidation of commands.
Documentation in the PDF (and the manual) is behind. I will do the PDF first, but it'll come after
all of this refactoring.