Currently investigating an issue with pel-goto-symbol.
It fails with some files, not all.  For example it fails in pel-benchmark.el.
The pel-goto-symbol-any-buffer which uses a different set of functions, is not affected.
When this fails, the issue is that imenu--make-index-alist does not identify valid markers for this file.
I'll look into this.  If anyone has seen this, please let me know.  Thanks.