Remove pel-toggle-imenu-index-follows-order and its user-option pel-imenu-index-follows-order-p.
This functionality is no longer needed as there are several packages that can be used with imenu
and these provide what is needed in order to see the declaration either in order or not.
It will be better to simplify and fully understand & document the interactions between the various
imenu backends, the semantic-mode, etc...  And easier to get this out of the way to fix issues
in imenu with its interaction with popup-switcher in some files or with the dynamic switching of the backend.