Restore ability to order imenu entries.  Add pel-imenu-toggle-auto-rescan, ..toggle-popup.
The issue related to imenu seems to be narrowing to popup-switcher.  So I restored the functionality I had removed.
Re-organized the commands and improved documentation both in the Menu and the navigation PDF.
More updates in the way info is displayed is coming.  Hopefully I'll be able to pin-point the popup-switcher bug.
It has another issue related to leaking "propertized" text.  Might be related.