Refactoring: use pel-setup-major-mode macro to configure each major mode. Shorter code, more complete behaviour.
The new logic ensures that file and directory variables are loaded when the mode control code runs.
This simplifies pel_keys code and fixed some code duplication.
It does introduce the ``<M-f12>`` key prefix in some major modes and I might remove it later if they're in the way.
Code is also a little faster.
Another change: added pel-use-c, pel-use-c++ used-options which control whether or not PEL logic for these languages
is implemented.  If they are not used, the startup will be a little faster.