Now take hide-lines from my Forked project until my PR is merged in.
Reason:
  The syslog-mode requires hide-lines and needs the last version of that code.
  The last version of the code was not available from MELPA because of a bug in MELPA spec.
  I reported that bug and it was fixed very soon after.
  However, hide-lines does not byte-compile cleanly.  I forked it and submit a PR to fix it
  back in January 17, but it has not need merged yet.

To update, the best way to proceed is to disable both pel-use-hide-lines and pel-use-syslog-mode,
run pel-cleanup, then re-activate both user-options and run pel-init.
If you do not proceed this way you might end up with a copy of hide-lines inside the elpa and the utils
directory.  Not a big deal because the one in utils will take precedence and that's the good one.  But
that will slow down emacs a little bit because of the extra useless directory inside elpa.