One more step toward a fast startup Emacs.
- Changed/simplified the strategy: instead of creating a directory with all
  one-level package emacs lisp files and create a set of package specs placed
  into the package-alist and fight Emacs package.el, now the code uses
  package.el logic: it still moves all one-level package Emacs Lisp files (and
  their byte compiled files) into a single directory, *but* that directory is
  the *fake* pel-bundle package.  The pel-bundle package directory is placed
  inside the elpa-reduced directory, along with the other multi-level packages
  that remain.  A pel-bundle-autoloads.el and a pel-bundle-pkg.el files are
  created to simulate a real Elpa compliant package, but its not really a
  package: it holds all one-level package Emacs Lisp files.
- The command that re-organize the files for the fast startup mode is now
  called: `pel-setup-bundled-operation-mode'.
- Add the `pel-restore-normal-operation-mode' command that restore the file
  organization the way they normally are, allowing Emacs to use the package.el
  mechanism to add or update external packages and PEL to automatically
  download and install them as well.
- Work is almost done but not yet complete: the code still need to handle removal of dependencies
  left in the -pkg.el files of the real multi-level packages inside the elpa-reduced
  directory.  Code comments must also be updated to reflect the new strategy.

Overall it looks like this strategy works and will reduce Emacs startup time a
lot when a large number of external packages are used.  I'll end up with 2
modes of operation: the normal Emacs mode of operation and the bundled mode
that provides a much faster startup but where installation of new packages
will not be supported.  It will be possible to switch from one mode to the
other as long as Emacs is restarted.

I have not yet tested it but I suspect it'll be possible to use other package
managers and use the same strategy.  For now that will be restricted to PEL.
But later I might learn enough to create a separate package that will allow
anyone to switch between a normal Emacs and an Emacs that has a faster startup
time.