pel-setup-fast and pel-setup-normal can't be used from Emacs in graphic mode and using package quickstart.
This is a limitation cause by PEL not being able to read the original value of the package-user-dir
and therefore not being able to see if the elpa directory is a symlink or not.
Either turn off package quickstart or use Emacs with quickstart running in the terminal to perform these commands.
Once the switch is made, Emacs in graphics mode using package quickstart is working fine.