Now getting fzf.el from my publish branch.
That branch contains code not yet integrated in main because I'm waiting to
get my changes integrated in the bling/fzf.el main.  This way it allows me to
provide new features to PEL users ahead bling/fzf.el.  And since I'm the main
user of PEL in various computers it allows me to easily distribute my latest
code.  Anyway it's as tested as what would be in my main branch.