Add function pel-minor-mode-auto-activated-by
It describes what PEL user-option requests activation
of a specific minor mode into a specific (or the current)
major mode.

By default the function acts as a predicate; it returns nil if nothing
requests its automatic activation. If the NIL-RETURN optional argument is
specified it can be made to always return a string (or something else), which
helps to use the function in message formatting generation.