Enhance pel-file-type-for and pel-minor-mode-auto-activated-by
`pel-file-type-for' now support an optional argument that can specify a prefix
to remove from a mode name.  By default it's "-mode" as before, but it can now
be used with "-minor-mode".

`pel-modes-activating-symbol-name-for' uses that to be able to identify minor
mode symbols that have a name that ends with '-minor-mode'.

Finally this is used by `pel-minor-mode-auto-activated-by'.