Fix: set tab-width upon visiting a buffer in one of the cc-modes.
`tab-with` is set to value specified `pel-XXX-tab-width` where XXX is the
major mode name (like c, c++, d).  Therefore the tab-width of a buffer in
c-mode is now locally controlled by the value of the `pel-c-tab-width`
user-option.

This was always planned but somehow I either removed the code or forgot to put
it in...