Associate .ii files to C++.  Update C++ tempo templates.
The C++ tempo templates now describe C++ files better.  They can distinguish
various types of implementation and header files from the file extension and
insert a better description inside the comment at the top of the file.

Also fix application of the `pel-X-tab-width` for D, Forth and Julia as it was
done for C and C++ in previous commits.