pel-abbrev-info buffer is placed in help-mode
This is the first of the many upcoming PEL info commands that will be using
the help-mode and will display information about all user-option variables
that have an impact on the behaviour of the feature described by the
information requested.

Some of the existing info command display a short message which shows up at
the bottom of the Emacs screen. Several have a lot more information to
display. I think I will convert all of these commands to display all the
relevant control variables and provide them as buttons, to help provide a
quick way to learn more about them, access their customization buffer, while
still printing them in a report fashion, with the name of the current buffer,
and a time stamp.  This allows the buffer's content to be used for report or
kept for comparison as the user experiments with the various features.