pel-cc-mode-info now uses a specialized help buffer
It also provides access to the help of relevant user-options and a quick way
to access their customization buffers.

Add pel-major-mode-symbol-for utility function.  This function as well as the
following listed below can use the pel-insert-symbol-content-context-buffer
contextual buffer overriding variable. The buffer argument of these function
overrides that variable, providing the following priority of the buffer
selection (first has highest priority):

- buffer argument
- pel-insert-symbol-content-context-buffer variable
- current buffer

The affected utility functions are:

- pel-major-mode-symbol-value-or
- pel-major-mode-symbol-value
- pel-major-mode-symbol-value