Improved PEL key bindings for iedit-mode.
- Improved PEL default key bindings for iedit-mode: by default PEL uses ~M-S-<f7>~, ~M-S-<f8>~ and ~M-S-<f9>~ to move to previous iedit occurrence,
toggle the iedit state of the occurrence and move to the next iedit occurrence, respectively, instead of the default ~backtab~, ~M-;~ and ~backtab~ keys.
This way, the standard behaviour of these often used keys can use used over an iedit occurrence.
The description of the iedit mode in the [[https://raw.githubusercontent.com/pierre-rouleau/pel/master/doc/pdf/search-replace.pdf][⅀  Search/Replace PDF]]  shows if a key binding is available over the iedit occurrence or anywhere when the iedit-mode is active.
- Add customization for iedit-mode: the new *pel-pkg-for-iedit* customization group holds the new
  *pel-iedit-use-alternate-keys* user-option that can be used to restore the default iedit key bindings.
  When text is highlighted by iedit-mode, the ~<f11><f2>~ and ~<f11><f3>~ open the  *pel-pkg-for-iedit* and *iedit*
  customization group buffer respectively.