Add support for root & side windows.  Prelim.  Window Hydra update.
Add pel-buffer-in-side-window to place current buffer in a side-window
identified by the numerical argument (2, 4, 6 or 8) corresponding to the
cardinal pointer positions of the numerical keyboard.

Updated the window Hydra to add those plus ability to create root windows at
the bottom and right side.

This is preliminary; I'd like to find a way to create root windows at the top
and the left.

And add support for atomic windows.

Not yest documented much.  Will experiment with what's here and then revisit.