Change Window Hydra keys for creating root windows to M-r and M-R
The use of the Meta key makes it more consistent.
It also leaves the 'r' for a more general resize that could mimic windresize
(or use it) it I found a way to redirect the cursor keys to it once it took
over.  For now use <f11> w r within the Hydra.