Improved key bindings used to  create side windows
The function pel-buffer-in-side-window is no longer a user facing command,
instead there's 4 new helper commands that do not require numerical arguments
to operate.  This simplifies their use and allows creating a more regular key
binding in both the global key bindings and reuses the implementation of
the Window Hydra which needed these functions anyway.

The new commands are:

- *pel-buf-in-side-win-top*,    bound to ~<f11> w \\ 8~ and to PEL Window Hydra ~\\ 8~,
- *pel-buf-in-side-win-bottom*, bound to ~<f11> w \\ 2~ and to PEL Window Hydra ~\\ 2~,
- *pel-buf-in-side-win-right*,  bound to ~<f11> w \\ 6~ and to PEL Window Hydra ~\\ 6~,
- *pel-buf-in-side-win-left*,   bound to ~<f11> w \\ 4~ and to PEL Window Hydra ~\\ 4~,