Add hydra support for window-purpose
Add a new hydra, the PEL window info hydra that is now the one
used to control the window dedication.  It supports dedicating a window,
and when window-purpose is available it can activate the purpose-mode
and use the functions to toggle the purpose-aware dedication of the window or
the buffer it holds.

Update the Windows PDF accordingly.  Will need to adjust inside links.