Add a new PDF that will be used to describe the display-buffer
display-buffer is an important function/variable to understand.
It's at the core of how Emacs open windows to display buffer.
The default behaviour is often annoying but it can be controlled
very flexibly with display-buffer once it's fully understood.

I'm accumulating links to important information and will try to
synthesize important aspects inside the new PDF, as I'm starting
to add support for it inside PEL to provide a better experience
under various circumstances.

The page is currently under construction.