Fix Emacs fast startup in Emacs 29.1
There is a problem remaining: the "Package autoload is deprecated"
message is still printed.
The switch to fast startup mode works and Emacs works
in that mode, but it complains about autoload package being deprecated.
The code change to not use it does not prevent the message from showing
and that message was printed last, preventing the user to see the
request to restart Emacs. I delayed the display of that message to ensure
that it is seen last.  However that does not solve the warning.

More investigation required.