ec help prints its limitation under macOS
The macOS implementation is currently not able to distinguish
an Emacs daemon that uses the default socket from one that uses
a named socket.

So if you start an Emacs daemon with a named socket, and later try to start
the Emacs daemon with the default socket, ec will fail, reporting it's already
running.

The work-around is to start the Emacs daemon with the default software
explicitly with the Emacs.app command.  Another way is to use ec to start the
Emacs daemon with the default socket first, and then start an Emacs daemon
with the name socket.