Improve navigation in sh-mode buffers.
Add commands to move point to next and previous function definitions.
Also add a command to toggle acceptance of hyphen and period inside function
names.

Some function names are user-facing inside the interactive shell as commands
and it's easier to type a command name that has embedded hyphens inside its
name than underscores.  They are valid in Bash and zsh anyway.   So might as
well promote the use of these characters inside the shell command names.

That's not done in the variable names.  They are normally not user facing.

Update the PDF and improved the order.