Removed compilation-minor-mode activation in shell/terminal
This does not work well and really, with PEL, it is *not* needed.
All that I need to use to open the file identified by the error is to use
the <M-f6> key binding ! That invokes the pel-open-at-point which does the
job.

That command might need a little bit more support of the shell/terminal modes
to have the ability to detect the location of the file, using a mechanism
similar to what is available in C/C++.  But it already works for valid
absolute or relative paths.