Tree-sitter activated for Python and TOML major modes
Using tree-sitter requires several changes.
All hooks and major mode related logic must be aware that the tree-sitter
major modes differ from the normal major modes, because of a decision in Emacs
to have them separate.

For PEL that means I have to duplicate some logic and ensure PEL supports the
2 sets of modes.  That forces creation of extra variables that copy the
content of the mode specific variables and have a -ts in their name.

Hooks must also be duplicated.

At this point I'm experimenting for the best way to support that
duplication. I want to eventually provide macros that will deal with that and
eliminate the dual variables.

For now I'm adding support for the toml-ts-mode and python-ts-mode
as an experiment.