Using new macro pel-config-major-mode-with-ts to simplify ts support
The new pel-config-major-mode-with-ts macro can now be used instead of
pel-config-major-mode to setup a mode that has a tree-sitter equivalent major
mode.

There are probably other aspects that need to be handled, like other hooks
created by other code.  But this is a step toward simplification and
elimination of code that would have had to be duplicated.