Preliminary tree-sitter support for C and C++
The code is not optimized, still has several loose ends that need testing.
The tree-sitter modes will be used by default when pel-use-tree-sitter is on.
If that causes a problem just turn it off and restart Emacs.

There's quite a bit of logic to go through in the case of C and C++ because
they are modes derived from cc-mode.  Lots of testing remains to be done to
really integrate tree-sitter support well in PEL.

Some functions are written with the assumption there's only one mode for C and
one for C++.  That's not the case anymore.