Improved integration with iflipb-kill-buffer
The C-x k key is now bound to iflipb-kill-buffer when iflipb is used.

The integration is not perfect because the code does not replace
kill-buffer everywhere.
For that I would need to modify the implementation of iflipb-kill-buffer
and use something like a defadvice.  This way, all invocations of kill-buffer
would update the iflipb data.