For HaraldJoerg/cperl-mode, use the upstream cperl-mode.el
Now get the upstream branch copy of cperl-mode.el from HaraldJoerg/cperl-mode
as well as the master branch copy of perl-tidy-ediff.el to get the best Perl
support.

The upstream branch copy of cperl-mode.el is the upstream copy of the official
Emacs cperl-mode.el that is used in the future version of Emacs (currently
Emacs 30) but is compatible with earlier versions of Emacs. This version
handles the Allman/BSD indentation style properly, unlike the master branch
copy and the current Emacs cperl-mode.el.

Updated PEL functions that interact with the Perl indentation functions of
cperl-mode.el accordingly.