pel-open-at-point handle filename followed by a shebang line
In some case, when the file name is followed by a file separator and a shebang
line the code now handles it properly.  That case happens when trying to
extract the file name from the names seen in a buffer that shows the output of
a grep search on files that output the first lines of files that have a
shebang line (for example). The code now handle this case.

Instead of trying to create a regexp that handles all cases, the code tries
several regexp.  The code now uses a cond form instead of a diff.  In future
if I find a new use case, I'll simply add a new clause to this cond form.