pel-perl-critic run remote host perlcritic when the file is remote
The code now checks if the file is remote and if it ensures that it checks the
presence of perlcritic on that same remote host and on it uses the local file
name from the remote host perspective.

Note: the logic to do all of this might better be located in perl--base where
it can be re-used in various locations, as well as it loads tramp dynamically.

Future optimization.