Fix generic tempo skeleton to handle presence of shebang line.
The code in pel-skels-generic now identifies the presence of a shebang line in
the buffer and places everything after the shebang line instead of above it as
it used to do it.

This way it becomes possible to execute `pel-as' in the buffer of a
extension-less file to set the shebang line and then add the file header
comment logic.