Better error message.
For PEL supported mode, the <f12> key is a prefix only when the
corresponding pel-use-MODE is set for the given major MODE.
When `pel-as' is invoked even if the major-mode is set, that may indicate
that the user did not activate the pel-use user-option.
The error message now describes this possibility.