Expand generation of skeletons to provide more help and sub-commands
Add ability to create a command that provides access to the mode's
skeleton configuration customization buffer.

Expanded `pel-tempo-install-pel-skel' docstring to provide all possible
information on how it is built.  Useful for debugging.

Add the <f12><f12> keymap for sh-mode buffers.  The keymap has a 'h' key
that creates the file header.   It uses the generic skeleton.
Eventually I might remove the <f6> key bindings because every supported mode
will have the <f12><f12> keymap.    This scheme belnds better with the use of
`pel-as'  bound to <f12> in fundamental-mode and no longer accessible once the
major mode is set and has a <f12> keymap.