Add pel-as and <f12><f12> prefix support for Python buffers.
It is now possible to create an executable extension-less Python file
by opening a file, typing <f12> in it to select python , which creates the
shebang line identified by the `pel-python-shebang-line' user-option, and then
typing <f12><f12>h to insert the file header.  Use <f12><f12><f2> to open the
customization buffer that defines the `pel-python-shebang-line' user-option.