Add Tcl & expect support for creating extension-less script files
The Tcl extension-less shebang line is controlled by the new user-option
`pel-tcl-shebang-line'.

The expect \ extension-less shebang line is controlled by the new user-option
`pel-tcl-expect-shebang-line'.

The <f12> key invokes `pel-as' which now fully supports Tcl and expect.
The <f12><f12>h key is supported in Tcl and expect, as much as the
<f12><f12><f2> key top access the customization buffer controlling the shebang
lines quickly.