Add preliminary Lua with support for creating extension-less script files
The Lua extension-less shebang line is controlled by the new user-option
`pel-lua-shebang-line'.

The Lua extension-less shebang line is controlled by the new user-option
`pel-lua-expect-shebang-line'.

The <f12> key invokes `pel-as' which now fully supports Lua and expect.
The <f12><f12>h key is supported in Lua and expect, as much as the
<f12><f12><f2> key top access the customization buffer controlling the shebang
lines quickly.

This is preliminary Lua support.  I have not yet pushed the PDF for Lua.