Fix bug: functions that interpret symbol: now do it in proper buffer
The following:
* `pel-symbol-on-off-string'
* `pel-symbol-value-or'
* `pel-option-mode-state'

have a new buffer argument and support the
`pel-insert-symbol-content-context-buffer' let-bound variable to be able to
report the state of a symbol in the proper buffer context if it is specified.

Also fixed all the functions that support the
`pel-insert-symbol-content-context-buffer' variable and an optional buffer
argument: in some cases neither is set; for those case the code now uses the
current-buffer.

This problem has been there for a while and affected several state reporting
commands.