No longer attempt to configure C and C++ tree-sitter mode
Emacs tree-sitter implementation has changed in Emacs 30.

It's better than what existed previously, but still not good enough for CC
derived modes: it is not as flexible to control the style and that is done in
a completely different way. The method I had used so far will never work for
the new tree-sitter support of Emacs.  So I removed all that code and
recommend to use c-mode and c++-mode instead of their tree-sitter equivalents.
Use tree-sitter only for rendering when you want to check something, but not
to write new code.

Later I might add more control for the tree-sitter but I will wait for the
evolution of Emacs' support first.