Cleanup: re-order like others.
I dropped the idea to put common code inside a source file.
POSIX shell scripts have too many limitations as far I know it today.
On the surface it look easy, but it's not because some common
code deals with the script arguments, and passing those to functions
defined in source scripts seems not possible.  I could pass the info via a
file, but that would cause other reliability problems for multi-processing.