Add pel-comment-dwim with supports commenting current line
pel-comment-dwim is now bound to M-; , replacing comment-dwim.
It is now possible to comment from the beginning of a line with M-0 M-;

That should be supported by comment-dwim but it's not.
comment-dwim code has a FIXME comment that should be addressed
by adding this feature IMO, but for now, I create this wrapper that does it.

Also update the C and Erlang specialized comment-dwim commands to provide the
same feature.