Add -n option to etags-c
This new flag allows the creation of a script that dynamically identifies the
names of the directories to analyze and use that command to get one directory
at a time, step-wise building a list of files to parse into a listfile without
building the TAGS file then.  And then build the TAGs file with another
command that uses the --list option.