pel-dirtree-find-replace fix: handle regexp new-text properly
One of the main reason to use a regexp in the new-text is to use matching
group in the text regexp to allow replacement of text surrounding the matching
group(s) without changing the matching group.

To be able to do that the string read from the user must be a string object
expressed in read syntax where backslashes are doubled-up.  The code now
treats the read specially and remind the user.

It might be possible to do all of the processing the way `replace-regexp' does
it and not impose the user to type the surrounding double quotes and use the
string syntax normally used in the prompt answers.  If I find how to do that I
will simplify the UI.