pel-dirtree-find-replace: new-text regexp is now string syntax
Improve the prompt capability of the new-text regexp entry: it is now entered
using the simpler string syntax instead of the read syntax. This means that a
match group only needs one backslash instead of two.

The code could be improved further by allowing the use of lisp expressions
that could be run in the replacement.  I might add this in the future once I
fully understand the way it is done in replace.el.