Better description of pel-show-indent
Describe what is shown, the buttons, the fact that customization values
correspond to what is used for new files but if you open a file written by
someone else the indentation scheme may differ from your configuration.  In
that case if pel-use-dtrt-indent is one the dtrt-indent-mode will be able to
detect the used indentation scheme and override the value of indentation control
variables with values that it has detected from the content of the file.