Add pel-indent-with-tabs-mode-for-dart
Setting pel-indent-with-tabs-mode-for-dart to a number tells PEL to
automatically activate pel-indent-with-tabs-mode minor mode with the specified
indentation width when opening Dart buffers.

This way it overrides the 2-space indentation used in the file with a
tabs-based indentation scheme used in the buffer and saves the buffer content
back to the file with the 2-space indentation scheme.  You can then conform to
the strict and inflexible Dart guideline while using something else inside
your buffer.