Better pel-open-file-alternate and better customization.
The new pel-open-file-alternate is able to detect more alternate files than
before; it has a better implementation that uses a mode-specific look-up
table.  The new implementation replace pel-alternate-extension-alist with the
more flexible pel-file-extension-alist.

pel-file-extension-alist identifies the list of code file extensions and
header file extensions for a specific major mode name.   This way,
pel-open-file-alternate can check if the current file is a code file or an
extension file and then search for a file that has the other extension, and
there can be several extensions for each.

The pel-file-extension-alist default value defines file extensions for Ada, C,
C++, Erlang and Seed7 as before but in a more flexible way, allowing you to
add more extensions for code and more extensions for header.

Since pel-file-extension-alist restricts the search for file extensions
specific to the major mode this prevents it from trying to open a file used
for another language.